Rename anchor component
