+++ src/components/Footer.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Container, Row, Col } from 'styled-bootstrap-grid';
 import styled from 'styled-components';
 import Text from './dls/Text/Text';
-import Anchor from './Anchor';
+import Anchor from './Anchor/Anchor';
 
 const FooterContainer = styled.footer`
   float: left;
